Replace all DMA_nBIT_MASK macro with DMA_BIT_MASK(n) From: Yang Hongyang <yanghy@cn.fujitsu.com> Date: Mon, 13 Apr 2009 14:40:14 -0700 Subject: [PATCH 1/5] Replace all DMA_nBIT_MASK macro with DMA_BIT_MASK(n)
This is the second go through of the old DMA_nBIT_MASK macro, and
there're not so many of them left,so I put them into one patch. I
hope this is the last round. After this the definition of the old
DMA_nBIT_MASK macro could be removed.

upstream commit e930438c42e744ef1f8bfdbb338253c9f384df42