b3dfg:  convert to triple buffering
Instead of allowing userland to drive the acceptence of interrupts from
the brontes board, continuously serve interrupts using a triple
buffering scheme.  This insures that userland can always lock one buffer
for use while the other two can be utilized to keep up with the most
recent triplets from the board.  Dropped triplets are now recorded when
we eject the oldest buffer in order to handle the newest.