b3dfg: restrict single user to buffers
The triple buffering scheme implemented in previous patches assumes that
only a single user will ever enable transmission and ask for buffers.
While this is how the hardware is currently being used, the ability to
toggle interrupts and lock buffers should still be restricted to one
application at a time.