b3dfg:  error on unhandled triplets from the hardware
With triple buffering, we should always be handling the interrupts from
the hardware, so no buffers should ever be dropped from the hardware's
perspective.  If the hardware reports any were, report an error.

Do not, however, update the triplets_dropped counter.  This is still
being used, but counts the number of triplets the driver has dropped due
to handling interrupts and being forced to drop the oldest populated
buffer in order to serve a dma request.