b3dfg:  enforce and track exclusive buffer locking
Userland should never possess a lock on more then one buffer at a time
as this can cause the driver to fall behind while servicing the
hardware.  As only a single buffer can be locked, it's easier for
everyone if the buffer is tracked by the driver and the user does not
have to supply the buffer number with release_buffer.