libb3dfg:  Remove set_transmission
The driver no longer requires a seperate call to set_transmission in
order to start receiving frames.  Instead, transmission is automatically
enabled on open and disabled on close.  Therefore we need to update the
library API in order to prepare the device handle in b3dfg_init() and
only open in b3dfg_open(). This also made the b3dfg_map functions
obsolete as b3dfg_get_buffer already returns the correct address.

Documentation and test cases have been updated as well to match this
change and represent the majority of the changeset.