b3dfg:  Drop first buffer after transmission is enabled
Frame transmission is handled by two separate pieces of hardware, the
framegrabber board (which this driver controls) and the wand connected
to the board.  The wand is controlled externally via an RS232 link.
In order for frame transmission to begin, the wand needs to be told to
start 'triggering' and the framegrabber needs to have transmission
enabled.  Typically this works correctly.

However, if triggering is enabled before transmission, the framegrabber
will partially fill the first buffer based on when transmission is
enabled.  The wand will then send the board a 'frame-complete' message
which the board happily accepts and then sends an interrupt without
verifying that it did indeed get an entire frame (hint, it did not).
This leads to a frame that is partially current and partially left over
from the last time triggering and transmission were enabled.

To workaround this issue, we make sure that we have already received
a second buffer from the framegrabber, therefore unconditionally ignoring
the first, potentially corrupt, buffer.