b3dfg:  reset DMA chip on start and stop
If something bad happens during a DMA transfer and the EC220_DMA chip
never gets an ack it will cause an interrupt storm where the DMA_STS is
never updated and the driver ignores the interrupts.  To prevent this,
we can simply reset the interrupt and error bits on the EC220 when
starting or stopping transmission.