b3dfg:  wait for debounce before checking cable state
We need to allow for at most 10ms for the cable status register to
debounce, so schedule a timer when we're alerted that the status has
changed and stop all interrupts until the state is known.

If userland requests a buffer during this period, they will fail due to
transmission being disabled.  However this is unlikely unless the cable
is bad and no worse than the previous behavior which could misread the
cable status a number of times and then deliver a bad first triplet.
Later patches should be able to improve upon the current behavior.