b3dfg:  check cable state while transmission is disabled
When transmission is disabled it is still important to track if a wand
is currently plugged into the system for thermal compensation.  This
patch enables the cable status interrupts at module load and after
handling a previous cable state change.

Note that due to a bug in earlier brontes board firmwares we need check
the cable state if an interrupt comes in while transmission is disabled.